Changed EEPROM reset script
