+++ platformio.ini
@@ -80,7 +80,7 @@ board_build.f_cpu = 24000000L
 build_flags = -D BURNWIRE
 lib_ldf_mode = deep
 
-[env:clear_eeprom]
+[env:reset_eeprom]
 platform = teensy
 board = teensy35
 framework = arduino
@@ -91,4 +91,4 @@ build_src_filter =
     -<**/*.cpp>
     -<**/*.h>
     -<**.*.enum>
-    +<../tools/clear_eeprom.cpp>
+    +<../tools/reset_eeprom.cpp>
